requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
sentence-transformers>=2.3.0
fastapi>=0.109.0
uvicorn>=0.27.0
qdrant-client>=1.7.0
python-multipart>=0.0.6
pypdf>=4.0.1
groq>=0.4.0
python-dotenv>=1.0.1
httpx>=0.26.0
pandas>=2.2.0
plotly>=5.18.0
pytest>=8.0.0
spacy>=3.7.0
rapidfuzz>=3.6.0
scikit-learn>=1.4.0
numpy>=1.26.0
pymongo>=4.6.0
python-jose[cryptography]>=3.3.0
